move GetDBTable to deps
